backend/requirement.txt
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.3
django-appconf==1.0.5
django-browser-reload==1.11.0
django-cors-headers==4.2.0
django-imagekit==4.1.0
django-ipware==5.0.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
executing==1.2.0
fonttools==4.43.0
h11==0.14.0
httpcore==0.17.3
httpie==3.2.2
httpx==0.24.1
idna==3.4
imageio==2.31.5
imutils==0.5.4
ipython==8.16.0
jedi==0.19.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
telebot==0.0.5
telegram==0.0.1
threadpoolctl==3.2.0
tifffile==2023.9.26
traitlets==5.10.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.8
